﻿[DataEntry:378] Copying row(s) of table data from one document into another is now disallowed.
[DataEntry:442]
Added feature to allow controls to be disabled and for the disabled status to be specified via a config file.  Added a new config file for AdvancedHealth that disables Epic patient name, MR Number and order number for phase 1.

[DataEntry:453]
Added "ValidationCorrectsCase" to pertinent controls so that it can be controlled whether a value that matches a validation list item case-insensitively will be modified to match the casing in the list.  The test/units column has this property turned off to preserve the casing used in the document.

[DataEntry:458]
Result date/time are no longer mandatory in the UI

Reviews pending.
[git-vault-id] Extract$/Engineering@867/868