﻿[DataEntry:400] Auto-update queries now have the ability to parameterize SQL queries to guarantee no special characters used in the parameters interfere with the query.
[DataEntry:403]
Added ability for an AutoUpdateQuery to be a "DefaultQuery" which fires only when the target attribute does not otherwise contain a value.

[DataEntry:481]
Multiple AutoUpdate queries are not supported

[DataEntry:482]
When the last pending document in a DataEntry validation task is processed, the document now closes instead of remaining open. Also, the DataEntry UI is not allowed to close when no document is running from the FAM since the DataEntry UI has no means in that case to inform the FAM of a close.

Also updated the DEP queries to reflect the new database schema and changed the font of the toolips to 13 instead of 15.

Reviewed by:

Nathan (2/2)
Jeff (3/6)
[git-vault-id] Extract$/Engineering@1114/1115