﻿[DataEntry:559] DataEntry framework is now licensed as DATA_ENTRY_COMPONENTS
[DataEntry:505]
The StandardLabDE DEP is now licensed as LABDE_VERIFICATION_UI_OBJECT.

[DataEntry:560]
Created [PFESFolder] tag that can be used in the config file to specify the Extract System directory.

[DataEntry:652]
Created a <SourceDocName/> auto-update query tag that can be used to populate a value with the current document name.

[DataEntry:563]
ApplicationTitle, ApplicationLogo and AboutLogo are now properties on the DEP.

[DataEntry:179]
Help menu & about box are now included.

[DataEntry:528]
Config files now includes DisabledValidationControls property to disable validation on specified controls.

[DataEntry:529]
The ordering physician table is now mandatory for the second reviewer only.

[DataEntry:534]
An image path and copy button are now include.  However, the copy button functionality is not yet implemented. Also, since the image path is a DataEntry control, it is not currently ideal to leave it enabled, but inactive so it is diabled. This results in the text being unselectable and not as readable. It would be nice to improve on this.

Reviewed by:

William (1/2)
Nathan (0/3)
[git-vault-id] Extract$/Engineering@1493/1494