﻿[DataEntry:578] Fixed a problem where tooltips could linger when they weren't supposed to.
[DataEntry:579]
Re-vamped the LabIdentifier auto-update query so that it tries several different ways to look up the lab, doesn't depend on the street address which AH wasn't able to provide, and so that it always assigns spatial information to the field for whatever is available from the name, city, state and zip (regardless of which fields were actually used to perform the lookup).

Also, fixed a problem where default queries would sometimes not execute if the attributes they were dependent upon were not yet available.

Reviewed by:
Jeff (2/0)
Wayne (0/0)
[git-vault-id] Extract$/Engineering@1662/1663