﻿[DataEntry:532] Fixed bug where tooltips that extend slightly off page cause an unnecessary zoom.
[DataEntry:572]
Added icon support to the DataEntryApplication and a LabDE icon to the LabDE DEP.

[DataEntry:581]
Fixed bug the marked table cells as viewed when a document was opened.

[DataEntry:588, 593]
Fixed exceptions an tooltip positioning issues that could occur when swiping text that was vertical (or nearly vertical) in the image viewer.

[DataEntry:601]
Tooltip are once again displayed for direct hints.

[DataEntry:602]
Fixed crash that occured when trying to commit a document when the copy button had not yet been pressed.

Also, some inefficiencies in query code had built up since the Beta2 timeframe that caused > 50% increase in load time of at least some documents.  I added optimizations to get the performance back to the Beta2 level.

Reviewed by:
Jeff (0/3)
Wayne (0/1)
[git-vault-id] Extract$/Engineering@1670/1671