﻿Fixed problem where pressing the copy button could cause an unhandled exception.
Fixed problem where unhandled exceptions could result in FAM mode from things like trying to write to a readonly file.

Adjusted a previous fix that prevented unexpected scroll and selection in tables so that it works for clicks on a row header as well as in a cell.

Reviewed by:
Jeff (1/1)
Nathan (0/0)
[git-vault-id] Extract$/Engineering@3714/3890