﻿$/Engineering/ReusableComponents/COMComponents/UCLIDFileProcessing/Code
FlexIDSCore #3717
Modified code to stop processing if after one thread is unable to get files if the stop processing once queue is empty setting is selected.  This will keep the situation were the queue is emptied for a thread then files are supplied and the threads that have not exited keep processing the newly supplied files.

LegacyRCAndUtils  5495
Added code to start and stop processing using a schedule.  As part of this hitting the pause button will now behave the same as processing inactive

Reviewed By
Jeff(1/8)
Nathan(0/2)

[git-vault-id] Extract$/Engineering@4589/4780