﻿Merge Branches: Origin=$/Branches/LabDE/1.5/Current/Engineering/ReusableComponents/BaseUtils/Code
Version: 27
$/Branches/LabDE/1.5/Current/Engineering/ReusableComponents/BaseUtils/Code

$/Engineering/ReusableComponents/BaseUtils/Code

LegacyRCAndUtils 5623 not a fix since it does not seem to be broken.
Added additional debug info so that an exception gets logged each time listening is started including restarts.  Also added the folder being listened to in the debug info for existing exception.

My testing of the FolderEventsListener did find that the listening did restart after a network disruption or computer reboot but the reconnection did take between 1 to 3 minutes before files dropped into the folder started to be seen.

Reviewed By
Nathan(0/0)
Jeff(0/0)

[git-vault-id] Extract$/Engineering@6218/6517