﻿Merged all LabDE 1.5 changes except build/install related changes back into the main branch.
NOTE: The LabDE solution has dependency issues in the main branch after merging... but it does build when the right projects are manually built first.

I am contemplating the fate of the LabDE solution now that the FlexDataEntry demo is going to use the DE framework... I will PVCS the build dependency issues if it is decided the LabDE sln will survive as a separate entitiy.
[git-vault-id] Extract$/Engineering@6392/6746