﻿$/Engineering/ProductDevelopment/AttributeFinder/Build
Fixed problem with build file not copying all of the installation files and folders.

For build 8.0.0.3 I manually copied the missing files so the install will work.  The Demo_FlexIndex and Demo_IDShield folders are mostly the same as 8.0.0.2. (there were no changes in Vault between versions for the demos)

William

[git-vault-id] Extract$/Engineering@6437/6791