﻿$/Engineering/ProductDevelopment/LabDE/Installation/DataEntry $/Engineering/ProductDevelopment/AttributeFinder/Installation/UCLID FlexIndex
FlexIDSCore #3921 - partial
Rearranged some files in the install moving them to FlexIndex merge module.

I still have a problem with InstallShield seeing a dependency with the DataEntry merge module and so it includes it in the IDShield install.  I have moved the only file that is not DataEntry related and put it in FlexIndex merge module so that only DataEntry related dll's are in the DataEntry merge module.

William

[git-vault-id] Extract$/Engineering@6452/6806