﻿$/Engineering/ProductDevelopment/AttributeFinder/Build
FLexIDSCore 3916
Modified make file to delete our merge modules from InstallShields merge module location.  This will keep InstallShield from adding a merge module that is out of date to an install if it believes there is a dependency.
This may not totally fix the problem in the SCR, but without the out of date merge module any merge module dependency that may be found by InstallShield will at least bring in up to date files.

William

[git-vault-id] Extract$/Engineering@6484/6838