﻿Added the initial draft for the FlexIndex demo using the DataEntry framework. In order to get the DataEntryTwoColumnTable to work (as it had been unused since early in the LabDE project), a number of fixes we necessary including a at least one already PVCS'd during LabDE 1.5. The process also exposed the cause of one previously un-reproducable SCR.
[DataEntry:765]
Found and fixed cause of previously un-reproduceable unhandled exception.

[DataEntry:861]
Validation objects now have 1 to 1 relationships with attributes (rather than with the controls that contain the attributes). This change was needed to get 2-column tables to work, but also solves the memory "leak" present in LabDE 1.5 and prevents problems that would have arisen in control configurations not present in AH's DEP.

[DataEntry:858]
Optimized by avoiding un-needed calls to SourceDocumentPathTags.Expand in the query system.

Reviewed by:
Nathan (2/3)
Jeff (0/3)
[git-vault-id] Extract$/Engineering@6643/6998