﻿$/Engineering/ProductDevelopment/Common $/Engineering/ProductDevelopment/AttributeFinder/Build $/Engineering/ProductDevelopment/LabDE/Installation/DataEntry
Made some changes Flex/IDShield build to create the DataEntry Merge module, moved some more items out of the merge module and into LabDE
I am not checking in changes to LabDE Install yet (still need to figure out a few things).  The DataEntry Merge module is will not actually be put in the installs since I need it built to be able to include it.

LegacyRCAndUtils #5676
Added GetFullUserName.exe to installs

William

[git-vault-id] Extract$/Engineering@6652/7007