﻿[FlexIDSCore:3959] Reworked ID Shield tester so that it generates all stats required for the hybrid workflow in one pass.
The ID Shield tester output is now structured as close as possible to the external statistics templates (though it will contain stats not present in the templates, especially in hybrid mode). The labels have also been changed to match the wording in the external templates.

[FlexIDSCore:3796]
The label for "Number of automated redactions" in hybrid mode is now more clear (both for the overall and automated stats)

[FlexIDSCore:3798]
The summary statistics file now reports both the DocTypesToVerify and DocTypesToRedact settings (when specified).

[FlexIDSCore:3910]
Limiting by type no longer results in an exception.

[FlexIDSCore:3912]
ID Shield tester now reports stats for the number of pages processed and the number of pages with expected redactions.

[FlexIDSCore:3980]
Though the UI previously allowed doc type condition to be specified for automatic redaction (DocTypesToRedact), only now is the setting honored.

Reviewed by:

Nathan (1/1)
Jeff (1/1)
[git-vault-id] Extract$/Engineering@6968/7343