﻿[LegacyRCAndUtils:5700]
SkipAuthenticationForServicesOnMachines DBInfo setting is now SkipAuthenticationOnMachines (schema 22 -> 23) which now avoids authentication regadless of the method an FPS file is run (FAM, service or RunFPSFile).

RunFPSFile now properly prompts for authentication either because of the "RequireAuthenticationBeforeRun" setting or because the FPS file is configured to run skipped files for all users.

The FAM Service now fails a FAM process that is configured to skip files for all users the same as it had previously done when user authentication was required.

Reviewed by:
Nathan (0/0)
William (0/0)
[git-vault-id] Extract$/Engineering@7307/7687