﻿I was mistaken in my original code for RunFPSFile in that I had thought the code I was using for processing against the database would result in the code paths being hit to mark a record as processing, update the FAST table, add an entry to the locked file table, etc.
I have now refactored the code to accomplish these tasks so that both normal FAM processing and RunFPSFile pass files being processed through the same code.

Also added the ability to attempt unlocking files when attempting to queue with force-processing a file that is in the processing state.

Reviewed by:
William (1/1)
Jeff (1/3)
[git-vault-id] Extract$/Engineering@7382/7763