﻿[DataEntry:891, 240] An old SCR dictating that data entry controls should use Verdana font was not working except in one isolated case in two-column tables which made it obvious there was a font issue. Tables now use Verdana as they should.
[DataEntry:886]
FlexIndes demo Legal description text box now has a scroll bar

[DataEntry:890]
Return address table rows no longer are resizeable.

[DataEntry:892]
Selection/deletion of multiple rows in the return address table is now possible.

Also made change so that when in edit mode in a two-column table and all contents of the cell are seleted, pressing delete deletes the spatial info of the cell as happens in a normal DataEntryTable.

Reviewed by:
Nathan (0/1)
Jeff (0/0)
[git-vault-id] Extract$/Engineering@7419/7803