﻿[DataEntry:3610]
Fixed two problems that was causing most of the automated test for SpatialProximityAS to fail.

One of the fixes was to Remove filtering of clues that was added 11/9/2009. At that time, the new verification task was leaving clues in the output so in order to have the RedactionTester return the results one would expect, clues had to be ignored in the expected VOA file. Clues are now being moved to metadata by the verification task unless they are turned on during redaction. Though it should be rare that clues are actually desired in the expected results, it should be allowed. Currently, the SpatialProximityAS automated test is configured in a way that expects clues to be included in the test.

Reviewed by:
Nathan (0/0)
William (0/0)
[git-vault-id] Extract$/Engineering@7429/7815