﻿$/Engineering/ProductDevelopment/AttributeFinder/AFSplitters/Code
FlexIDSCore #4097
Updated the code that was checking the attributes returned from the MuniSplit.rsd file for the Old attribute names that contained a space that causes the Municipality subattribute to not be output to the results. Updated the NRS expected EAV files to change the expected type to same as in the MuniSplit.rsd file.

Reviewed By
Nathan(0/0)

William

[git-vault-id] Extract$/Engineering@7663/8051