﻿$/Engineering/ReusableComponents/COMComponents/UCLIDFileProcessing/Code
LegacyRCAndUtils #5741
Modified behavior of the stop button in the FAM UI so that cancelable tasks will stop immediately, even in the case when a pause is resumed before the processing paused and the stop button is pressed.

FlexIDSCore #3983
Fixed problem of the UI getting out of sync with the processing when using the Pause button.  Also if stop is pressed all buttons and menu items will be disabled  until processing stops

Reviewed By
Nathan(0/1)

[git-vault-id] Extract$/Engineering@7797/8189