﻿[DataEntry:889] Fixed a problem that was preventing the delete key from deleting a table row.
[DataEntry:4062]
The FlexIndex demo DEP now has a unified party table with "Party Type" and "Mapped To" columns. The party types are validated.

Added DocumentType/PartyType data to the database to use to populate possible document/party type lists and to validate a valid party types for the current document type. Also changed "Saint Louis" to "St Louis" and "Saint Paul" to "St Paul" in the zip code table to match the text on the demo documents.

Also fixed a bug that could cause infinite recursion in a table wben adding a new row if a column's validation rule doesn't allow and empty value.

Reviewed by:
Nathan (0/0)
[git-vault-id] Extract$/Engineering@7883/8302