﻿$/Engineering/ReusableComponents/COMComponents/UCLIDFileProcessing/Code $/Engineering/ReusableComponents/COMComponents/UCLIDFileProcessing/FAMUtils/Code
DotNetRCAndUtils #419, 420 LegacyRCAndUtils #5750, 5500
Modified the code to set a minimum auto revert time out of 5 minutes so that if it is less than 5 the value will be updated in the DBInfo table. Also changed code to call the pingDB method before checking for files to revert so that connection retries will not cause the auto revert to be triggered.

Reviewed By
Jeff(1/2)

[git-vault-id] Extract$/Engineering@7899/8318