﻿[LegacyRCAndUtils:5032] Arrow keys no longer allow scrolling while in an SRIRImageViewer rubberband operation.
[LegacyRCAndUtils:5258]
Made changes to reduce the memory usage from queuing a large number of added files.

SCR NOTE: The memory savings are not what I imagined they could be when I created the SCR (looks like possibly as little as 25% from the testing I have done so far) partially because I did not make all the changes implied by the SCR. It was not practical to use filename filtering to prevent dispatching events from FolderEventsListener because the class was not really designed to be used that way and there was also some concern that too much processing in the listening loop could cause files to be missed (I think more than likely this would not have been the case, but it doesn't make sense to risk it). Anyway, the scenario being handled by filtering event dispatches on filename would be an uncommon situation in the real world anyway: that non-image files were being added alongside the images that were to be processed.

Reviewed by:
William (0/2)
[git-vault-id] Extract$/Engineering@7962/8381