﻿[DataEntry:910] DataEntryComboBoxCells now correctly display error icons. The error text of the MappedTo field has been changed to be more specific.
[DataEntry:911]
Fixed behavior where deleting the city or state of the return address would cause the fields to be repopulated after tabbing out of the cell via an auto-update query using the zip code. The fix utilizes a new "TriggerUpdate" attribute that can be set on an attribute query node to indicate whether changes to a particular attribute will cause the query to update.

[DataEntry:912]
Fixed issue where swiping into a two-column table would not update a cell in edit mode.

[DataEntry:913]
Fixed error that could be generated when swiping a value into a ComboBox table cell that matched a possible value case-insensitively but not case-sensitively.

[DataEntry:915]
Help file link is now correct.

[DataEntry:916]
Changed "FlexIndex" references in the demo DEP to "FLEX Index".

Reviewed by:
Nathan (0/0)
[git-vault-id] Extract$/Engineering@7990/8409