﻿[FlexIDSCore:3613, 3729]
Fixed issues with the eav files for the OCRImageArea automated test:

OCRArea1.eav: Did not seem to correspond at all with the rule's settings. The output appeared correct, so I used it to create a new eav file.

OCRArea2.eav: This test involves a lot of text and will probably change every time the OCR engine does. The output appeared correct, so I used it to create a new eav file.

There is a legitimate issue that exists with OCRArea3. This has been PVCS'd as [FlexIDSCore:4155], but not tagged for 8.0.
[git-vault-id] Extract$/Engineering@8016/8444