﻿Merged changes from LabDE 1.5 patches into the main code branch. Confirmed behavior in 8.0 and that no crashes are observed when running with gflags.
[DataEntry:932, 957]
Confirmed these validation issues relating to [DataEntry:957] are fixed in the the current baseline.

When Nathan was trying to reproduce another manifestation of validation problem, he was unable to but his attempt did trigger a "COM object that has been separated from its underlying RCW cannot be used." exception when closing the DataEntryApplication that I will be SCR'ing separately.

Reviewed by:
Nathan (1/0)
[git-vault-id] Extract$/Engineering@8919/9465