﻿$/Engineering/ReusableComponents/COMComponents/UCLIDCOMUtils/Core/Code $/Engineering/ProductDevelopment/AttributeFinder/AFOutputHandlers/Code
FlexIDSCore #3658
Fixed the behavior of the RunObjectOnQuery output hander so that if an output handler is the object ran any attributes that are added or removed from the root level by the RunObjectOnQuery output handler will be reflected in the attributes output.

LegacyRCAndUtils #5833
Fixed problem in IUnknownVector::RemoveValue that will cause indeterminate behavior when an attribute was in the vector more that once.

Reviewed By
Jeff(0/2)

William

[git-vault-id] Extract$/Engineering@8933/9486