﻿[DataEntry:928] DataEntry input event tracking and counters now work in the current codebase.
Added a property to the Highlight class that allows the code that ensures WYSIWYG compatiblity with the SportIRW to be turned off. This code is expensive when there is a large number of highlights being created (it added about a second to the load time of a medium/largish LabDE document) and it the WYSIWYG compatibility is not needed for DataEntry.

Also, my previous checkin had opened up a code path that caused another RCW error from accessing a finalized attribute after swithing documents in stand-alone mode. In addition to solving the specific situation, I re-ordered DataEntryControlHost.ClearData to prevent any similar situations from happening.

Reviewed by:
Jeff (1/0)
[git-vault-id] Extract$/Engineering@8963/9530