﻿$/Engineering/ProductDevelopment/AttributeFinder/Build $/Engineering/ProductDevelopment/Common/ExtractCommon $/Engineering/ReusableComponents/VendorSpecificUtils/SafeNetUtils/Installation/Extract Systems LM $/Engineering/ProductDevelopment/AttributeFinder/Installation/UCLID FlexIndex
Related to
FlexIDSCore #551,2558,2972,3460
Attempting to fix the various Uninstall issues where not all files are removed properly.  I have created a new merge module that has the files that are common to Extract Systems LM and the UCLID Flexindex merge module, I have eliminated the compile of LandRecords.sln file and the InputFunnel merge module and moved any files that were installed with the InputFunnel to UCLID FlexIndex merge module.

It make take me several builds of the Install to verify that everything is in the correct locations for the building of the installs.

William

[git-vault-id] Extract$/Engineering@10285/10971