﻿[DataEntry:998], [FlexIDSCore:4293]
When the LabDEOrderMapper detects an updated database, when copying the file to the local temporary location it now uses a new file name so that other OrderMapper instances which are still referencing the old temp file are not disrupted.

If CacheRSD is on, the old temp files will be removed after each OrderMapper instance switches to the new DB copy. With CacheRSD off, the old copies will remain until the FAM is closed (triggering the dispose call on all of the OrderMapper instances that were created).

Reviewed by:
Jeff (2/1)

[git-vault-id] Extract$/Engineering@10441/11127