﻿$/Engineering/ReusableComponents/COMComponents/UCLIDFileProcessing/Code $/Engineering/ReusableComponents/COMComponents/UCLIDFileProcessing/Utils/FAMDBAdmin/Code
LegacyRCAndUtils #5866
Modified code to use FileActionStatus table and no longer create Action column in FAMFile table.

The use of the FileActionStatus table is determined by the value of
HKCU\Software\Extract Systems\ReusableComponents\COMComponents\UCLIDFileProcesing\FileProcessingDB\UsePreNormalized

registry key.  If it is 1 the FileActionStatus table will be used and the database schema will be 100 if it is 0 the database will use schema 23 (same as 8.0)

LegacyRCAndUtils #5884
Fixed problem with export list throwing an exception if selecting with a Tag related option.

Reviewed by
Jeff(2/4)

William

[git-vault-id] Extract$/Engineering@11199/11887