﻿[FlexIDSCore:4450] Added support for full screen mode in the redaction verification task. F11 can be used to toggle full screen mode, but there is also a configuration property which will force the verification window to be opened in full screen mode by default.
As part of this change, I replaced the FormMemento class and & IControlMemento interface with a more extensible, resuable and easier-to-use FormStateManager class that now includes a lot of code that was previously repeated in FormMemento users. The FormStateManager has a FullScreen property to make it easy to support full screen mode in any form and to persist whether or not full screen mode is being used between sessions.

Reviewed by:
William (2/1)

[git-vault-id] Extract$/Engineering@12457/13182