﻿[FlexIDSCore:3394] MergeAttributes will now take page skew into account when creating merged regions. Therefore, merged regions on skewed document should now be much more appropriate than in the original version of the rule.
[FlexIDSCore:4381]
The MergeAttributes rule now has a new output option that specifies whether the merged attribute will be created by merging the overall attribute bounds or the individual raster zones of the attributes. This enables the situation to be avoided where an attribute that wraps lines will have an unexpectedly large region wherein attributes will be merged.

- MergeAttribute automated tests were not being called as part of overall automated tests. Now they are.
- Corrected MergeAttributes automated tests so that they no longer use the depricated LegalDescriptionFinder
- Corrected MegerAttributes automated test to account for the previous change in overlap claculation (FlexIDSCore:3509)
- Added automated tests for FlexIDSCore:4381 and the new output option.

Reviewed by:
Jeff (1/2)
[git-vault-id] Extract$/Engineering@12541/13266