﻿PD\AF\IS\Redaction\Utils\IDShieldStatisticsReporter RC.Net\Utilities
[FlexIDSCore:4486]

Added a new checkbox should be added on the Analyze tab of the statistics reporter to allow manual data items to be considered automatically redacted.

I also needed to fix a bug I introduced in a recent checkin of ConfigSettings that prevented the statistics reported config file from being loaded.

Review pending...
[git-vault-id] Extract$/Engineering@12993/13734