﻿[FlexIDSCore:4514] Redaction fill color is no longer cut off in magnifier window when the magnifier is bigger than the image viewer.
[FlexIDSCore:4520]
Tracking rectangles and redaction fill color in the magnifier window are  now clipped at the edge of the image bounds to mirror what happens in the image viewer.

[FlexIDSCore:4524]
The image viewer will no longer inappropriately scroll in circumstances where the magnifier is active but the image viewer has a higher zoom than the magnifier.

Reviewed by:
William (0/0)

[git-vault-id] Extract$/Engineering@13451/14244