﻿[FlexIDSCore:4495] Fixed issue where a very small selection with the word highlight tool could end up highlighting a word that contained the upper-left corner (coordinate 0,0) of the image viewer control.
[FlexIDSCore:4535]
Fixed redaction line-fitting malfunction.

I have noticed the magnifier is a CPU hog. I’m going to make a separate SCR to look into ways to make it less of a drain on CPU, but for the time being I have added code to ensure it is not magnifying while changing images (otherwise if you moved the mouse while changing images it could dramatically increase the time it took for the next image to load.

Jeff (1/0)

[git-vault-id] Extract$/Engineering@13615/14419