﻿[FlexIDSCore:4483] Fixed casing of slideshow tooltips.
[FlexIDSCore:4596]
In order to prevent the slideshow from continuing to run beyond the last file in the queue, if a document comes up for verification more than 5 seconds after the last document closed, the slideshow will be automatically stopped.

[FlexIDSCore:4607]
Added progress indicator bar across the top of the image viewer to make it clear when the slideshow is running.

Reviewed by:
Jeff (2/1)

[git-vault-id] Extract$/Engineering@14185/14991