﻿[FlexIDSCore:4604] Clues are now merged and maintained by the VOA merge and compare tasks (note: all turned-off attributes will appear under the  “OldRevisions” node, but will still be displayed if the file is re-verified).
[FlexIDSCore:4605]
The VOA file merge and compare tasks now fail if at least one of the source VOA files is missing.

[FlexIDSCore:4620]
If merging two VOA files where one of the VOA files was the result of a previous merge, the metadata from the first merge is now properly nested in the output of the second merge.

[FlexIDSCore:4623]
The VOA file merge/compare tasks now preserve and merge non-spatial attributes such as DocumentType.

Also did some commenting/cleanup in the WordHighlightManager to help prevent against the introduction of any threading issues.

Reviewed by:
Jeff (1/4)

[git-vault-id] Extract$/Engineering@14425/15231