﻿[FlexIDSCore:4633] [FlexIDSCore:4634] Implemented seamless navigation which allows page and tile navigation to seamlessly move between documents. Tab and Shift+Tab will now have the ability to move both forward and back between documents whether or not seamless navigation is enabled. As part of this change, zoom history will be maintained for pages even after switching between documents and, likewise, when using tile navigation between documents in seamless mode, the same zoom factor will be maintained across document boundaries (though if the images are different size or resolutions, the zoom may appear different). In general, with seamless mode active (and prompts for not yet committed documents off) the only difference between navigating one document with thousands of pages and thousands of one page documents should be: - The document name will change in the title bar. - The first and last page buttons/shortcuts will operate within the current document. - Tags and comments will continue to change on document boundaries.
Also, though thumbnails will not currently be displayed for other documents, that may be something to consider in the future in keeping with the “seamless” concept.

[FlexIDSCore:4636]
Implemented “Prompt to save changes before navigating away from uncommitted documents”  option so that, by default, users will no longer be prompted before committing documents for the first time even if they are not being explicitly committed (ie, tab key or page down, next tile in seamless mode).

[FlexIDSCore:4649]
Corrected misspelling in the UI checkbox for the above option.

Reviewed by:
Jeff (2/7)

[git-vault-id] Extract$/Engineering@14713/15530