﻿I'm rolling back the optimization experiment. Joanna noticed some automated tests involving LIR were failing that hadn't before. I found that on my machine in release mode when I ran these tests it crashed in the same point in the same licensing call where I was seeing the FlexIDSCore:3073 crash.
I strongly believe that this is a problem in our code and not an error in the visual studio optimization, but despite the consistency of the latest incarnation of the crash (as well as the FlexIDSCore:3073 version) my efforts to uncover the true cause have stalled and I cannot justify spending more time on it (at least for the time being).

For the now at least, I am leaving the change to make the LicenseManagment class static since Jeff is working in that code. We can discuss whether to roll that change back as well once Arvind returns.

Reviewed by:
Jeff (0/0)

[git-vault-id] Extract$/Engineering@14829/15646