﻿[LegacyRCAndUtils:6082] The doc tags menu now includes a parameter list for doc tag functions.
[LegacyRCAndUtils:6033]
Also, on the c++ side I unified the call to display the doc tag menu that used to be accomplished via more than a half a dozen separate project-specific classes. As part of this unification, the c++ path tags button will now set the cursor between the parens of an added path-tags button just as the .Net path tag buttons do. Additionally, with both the .Net and c++ path tag buttons, if there is selected text when a path tag function is inserted, the selected text will be replaced, but then re-inserted within the function’s parentheses.

Reviewed by:
William (0/0)

[git-vault-id] Extract$/Engineering@15013/15851