﻿[LegacyRCAndUtils:5966] Added application trace log when upgrading schema.
[LegacyRCAndUtils:6088]
Fixed typo in database options dialog

[LegacyRCAndUtils:6089]
All exceptions that are displayed to the user will now be prefixed with “Displayed:” in the exception log. There is a new option “Show displayed exceptions” in the UEXViewer which allows these exceptions to be filtered out.

[LegacyRCAndUtils:6090]
I reviewed exceptions looking for ones that should be “Application Traces”. The following exceptions became application traces:

...

I made the following exception not an application trace:
ELI24024 File cannot be opened for reading.

Also, the exception that logs the USB counts has changed to “Application trace: debug information”.

Reviewed by:
William (2/2)

[git-vault-id] Extract$/Engineering@15026/15864