﻿$/Engineering/RC.Net/FileActionManager/FileProcessors/Core/Code $/Engineering/RC.Net/FileActionManager/FileSuppliers/Core/Code $/Engineering/RC.Net/Utilities/Core/Code
DotNetRCAndUtils #677
FtpFileSupplier now creates a .info file when a file is downloaded.
The .info is in XML format and contains values for
    RemoteSourceDocName - full path of file on ftp server
    SourceDocSize - size of file on ftp server
    LastModifiedTime  - last modified time of file on ftp server

DotNetRCAndUtils #678
FtpFileSupplier checks for the file and an associated .info file before downloading. if the last modified time and file size are the same the file is not downloaded

DotNetRCAndUtils #679
Added support for the <RemoteSourceDocName> to the  FtpFileTransferTask

Reviewed By
Steve(1/1)

William

[git-vault-id] Extract$/Engineering@15399/16441