﻿[FlexIDSCore:4565] I have reworked redaction resizing: - There is no longer any distinction between angular and rectangular redactions as far as resizing goes. All 8 grip points are now visible and useable on all redactions. - The resizing is no longer forced to “rotation” mode for the left and right grip handles of angled redactions. - All sizes can now be moved individually. There is still some wobble, but it should be within reason. (I’ll SCR this separately).
[FlexIDSCore:4105, 4166]
Fixed issue where redaction side jumps when a grip is initially clicked.

[FlexIDSCore:4562]
As part of this,  have reintroduced the ability for ‘+’ and ‘–‘ to affect all sides of a redaction (not just the height). The angle should now remain constant, however there is still a lesser issue remaining in that a redaction that is enlarged & shrunk repeatedly will gradually change its “aspect ratio” (width/height). I’ll SCR that separately, but it may be difficult to solve.

[FlexIDSCore:4541]
Though the aspect ratio can still change, there should not be any situations where redactions stop shrinking either in the horizontal or vertical dimension.

Reviewed by:
William (0/2)
[git-vault-id] Extract$/Engineering@15546/16590