﻿The Extract.Imaging.Forms.Highlight and Extract.Imaging. RasterZone classes have been modified so that they now use floating point coordinates rather than integer coordinates. While the coordinates are still persisted to disk as integers for compatibility with COM RasterZones, by rounding off only at the end of a tracking operation rather than during it,  the visually distracting wobbling/bobbling that had still occurred when resizing highlights in build 35 has now been solved.
Highlights will often need to make a slight adjustment in size/position when the mouse is released, however… but the adjustment should be very small and generally mistakable for a slight movement the user may have made when releasing the mouse button.

Also, I have tweaked the order in which painting and tracking updates occur when the magnifier is active—it turns out that during tracking events the user was actually seeing that state of the highlight following the previous MouseMove event rather than the current MouseMove event.

[FlexIDSCore:4724]
Temporary highlight resize logging has been removed.

Reviewed by:
William (0/1)

[git-vault-id] Extract$/Engineering@15580/16632