﻿[LegacyRCAndUtils: 5902, 6036, 6037, 6038, 6039, 6040, 6041] Fixed issues where the lack of ability to create default values in HKEY_LOCAL_MACHINE caused exceptions when running without necessary permissions. Any code that fails to create a registry key for the purpose of installing a default value will ignore the exception unless the key was being created in the HKEY_CURRENT_USER hive. Note the USBLicenseKey manager still writes to HKEY_LOCAL_MACHINE but will now request elevated permissions when it is run.
[FlexIDSCore:1285]
Optimized the document classifier so that it doesn’t perform unnecessary searches. In the example case, this led to 5x better performance.

[FlexIDSCore:4719]
Fixed Select Page Region on pages so that regions don’t get cut off on pages with rotated orientation.

[FlexIDSCore:2587]
When splitting attributes using an RSD file, it is no longer necessary to re-run document classifiers that were run in the parent rule.

[FlexIDSCore:3590]
Implemented “FindXOf” block for document classifiers. In such a block the specified number of patterns must match for the block to be considered a match.

Reviewed by:
William (1/2)

[git-vault-id] Extract$/Engineering@15744/16828