﻿[FlexIDSCore:4655] The ModifyVOA utility now modifies evoa files as well as voa files.
[FlexIDSCore:2525]
Where LIR always used to base the boundary offset on average character/line width/height of the clue for boundaries based on a clue and the average character/line width/height of the page for boundaries based on the page, 2 or 3 options are now offered.

For boundaries based on the page bounds:
“Inches”
“Page Lines/Characters” (average size of the characters/lines on the page)

For boundaries based on a clue:
“Inches”
“Clue Lines/Characters” (average size of the characters/lines of the clue)
“Page Lines/Characters”

[FlexIDSCore:4590]
Unrecognized characters included in OCR results are now assigned a confidence of 1 rather than 229.

Reviewed by:
William (0/0)

[git-vault-id] Extract$/Engineering@15791/16875